movies_admin/requirements.txt
asgiref==3.8.1
autopep8==2.1.0
black==24.3.0
certifi==2024.2.2
click==8.1.7
Django==5.0.4
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-split-settings==1.3.1
djangorestframework==3.15.1
elastic-transport==8.13.0
elasticsearch==8.13.0
elasticsearch-dsl==8.13.0
flake8==7.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
uWSGI==2.0.25
